requirements.txt
defusedxml==0.7.1
fonttools==4.47.2
fpdf==1.7.2
fpdf2==2.7.7
Levenshtein==0.23.0
pillow==10.2.0
python-barcode==0.15.1
python-Levenshtein==0.23.0
rapidfuzz==3.6.1
